+++ src/proxy/shield.go
@@ -7,7 +7,6 @@ import (
 	"net/http"
 	"fmt"
 	"net"
-	"os"
 	"math"
 	"strconv"
 )
@@ -250,8 +249,9 @@ func calculateLowestExhaustedPercentage(policy utils.SmartShieldPolicy, userCons
 
 func GetClientID(r *http.Request) string {
 	// when using Docker we need to get the real IP
+	UseForwardedFor := utils.GetMainConfig().HTTPConfig.UseForwardedFor
 
-	if os.Getenv("HOSTNAME") != "" && r.Header.Get("x-forwarded-for") != "" {
+	if UseForwardedFor && r.Header.Get("x-forwarded-for") != "" {
 		ip, _, _ := net.SplitHostPort(r.Header.Get("x-forwarded-for"))
 		utils.Debug("SmartShield: Getting client ID " + ip)
 		return ip