+++ src/utils/types.go
@@ -124,6 +124,7 @@ type HTTPConfig struct {
 	OverrideWildcardDomains string `validate:"omitempty,excludesall=/ "`
 	AcceptAllInsecureHostname bool
 	DNSChallengeConfig map[string]string `json:"DNSChallengeConfig,omitempty"`
+	UseForwardedFor bool
 } 
 
 const (